+++ CHANGELOG.md
@@ -13,7 +13,8 @@
 - added `getSingleTax` method to `Cart\CartItem`
 - made `getSinglePrice` and `getTotalPrice` methods on `Cart\CartItem` always return price including tax
 - added `getSinglePriceExlcudingTax` and `getTotalPriceExcludingTax` methods on `Cart\CartItem`
-
+- made `total` method on `Cart\Cart` always return total including tax
+- added `totalExcludingTax` method on `Cart\Cart`
 
 ##2.0.0